+++ sql/core/src/main/scala/org/apache/spark/sql/DataFrameReader.scala
@@ -197,7 +197,7 @@ class DataFrameReader private[sql](sparkSession: SparkSession) extends Logging {
     val lookupCls = DataSource.lookupDataSource(source, sparkSession.sessionState.conf)
     val cls = lookupCls.newInstance() match {
       case f: FileDataSourceV2 if disabledV2Readers.contains(f.shortName()) ||
-        disabledV2Readers.contains(lookupCls.getCanonicalName) =>
+        disabledV2Readers.contains(lookupCls.getCanonicalName.toLowerCase(Locale.ROOT)) =>
         f.fallBackFileFormat
       case _ => lookupCls
     }
