+++ sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/FallbackFileDataSourceToV1.scala
@@ -26,10 +26,10 @@ import org.apache.spark.sql.execution.datasources.v2.{DataSourceV2Relation, File
  * Replace the V2 data source of table in [[InsertIntoTable]] to V1 [[FileFormat]].
  * E.g, with temporary view `t` using [[FileDataSourceV2]], inserting into  view `t` fails
  * since there is no correspoding physical plan.
- * This is a temporary hack for making current data source V2 work. It should be removed
- * when write path of file data source v2 is finished.
+ * SPARK-23817: This is a temporary hack for making current data source V2 work. It should be
+ * removed when write path of file data source v2 is finished.
  */
-class FallBackFileDataSourceToV1(sparkSession: SparkSession) extends Rule[LogicalPlan] {
+class FallbackFileDataSourceToV1(sparkSession: SparkSession) extends Rule[LogicalPlan] {
   override def apply(plan: LogicalPlan): LogicalPlan = plan resolveOperators {
     case i @ InsertIntoTable(d @
         DataSourceV2Relation(source: FileDataSourceV2, table: FileTable, _, _, _), _, _, _, _) =>
